Add documentation for ALIAS / PERF LAB
This clones the scale lab "bm" and "sno" readme documents into "scale" and
"alias" versions, making adjustments for ALIAS setup.

I renamed the set of documents to make them more consistent.

I've probably missed stuff, and suggestions or corrections are welcome, but I
figured that putting it up early is better.

(Also, while I kept this "simple" by cloning separate ALIAS documents, there
aren't that many differences and maybe it'd be better to combine the ALIAS and
SCALE documents with multiple tables as in `tips-and-vars.md`.)

Resolves #313